BackgroundLogWriter: triage remaining policy findings
- Document the BackgroundLogLineSink.write requirement (DocC).
- Replace the private BackgroundLogEntry struct with a private tuple typealias so
  the writer file holds a single major type.
- Move the test RecordingSink double into its own file (test-double convention).

The remaining FileManager.default use is in FileBackgroundLogLineSink, the
internal concrete adapter behind the injectable BackgroundLogLineSink seam — the
one correct place for real filesystem access, and it is covered directly by
fileSinkAppendsThroughOneLongLivedHandle against a temp file.
